+++ src/test/groovy/edu/ie3/simona/api/simulation/ExtSimulationSpec.groovy
@@ -1,6 +1,7 @@
 package edu.ie3.simona.api.simulation
 
-import edu.ie3.simona.api.data.ExtSimAdapterData
+import edu.ie3.simona.api.data.SetupData
+import edu.ie3.simona.api.data.connection.ExtSimDataConnection
 import edu.ie3.simona.api.data.connection.ExtDataConnection
 import edu.ie3.simona.api.ontology.simulation.ActivationMessage
 import edu.ie3.simona.api.ontology.simulation.CompletionMessage
@@ -71,9 +72,10 @@ class ExtSimulationSpec extends Specification {
         def tick = -1L
         def newTick = 0L
         def extSimAdapter = testKit.createTestProbe(ControlResponseMessageFromExt)
-        def extSimData = new ExtSimAdapterData(extSimAdapter.ref(), new String[0], null, null)
+        def extSimData = new ExtSimDataConnection(extSimAdapter.ref())
         def extSim = new TestSimulation(newTick, Optional.of(-2L))
-        extSim.setAdapterData(extSimData)
+        extSim.setDataConnection(extSimData)
+        extSim.setSetupData(new SetupData(new String[0], null, null))
 
         when:
         extSimData.queueExtMsg(new ActivationMessage(tick))
@@ -87,11 +89,12 @@ class ExtSimulationSpec extends Specification {
     def "An ExtSimulation should handle activation and return given new triggers"() {
         given:
         def extSimAdapter = testKit.createTestProbe(ControlResponseMessageFromExt)
-        def extSimData = new ExtSimAdapterData(extSimAdapter.ref(), new String[0], null, null)
+        def extSimData = new ExtSimDataConnection(extSimAdapter.ref())
         def newTickOpt = newTick.isEmpty() ?
                     Optional.<Long>empty() : Optional.of(newTick.first())
         def extSim = new TestSimulation(-2L, newTickOpt)
-        extSim.setAdapterData(extSimData)
+        extSim.setDataConnection(extSimData)
+        extSim.setSetupData(new SetupData(new String[0], null, null))
 
         when:
         extSimData.queueExtMsg(new ActivationMessage(tick))
@@ -112,9 +115,10 @@ class ExtSimulationSpec extends Specification {
     def "An ExtSimulation should handle termination properly"() {
         given:
         def extSimAdapter = testKit.createTestProbe(ControlResponseMessageFromExt)
-        def extSimData = new ExtSimAdapterData(extSimAdapter.ref(), new String[0], null, null)
+        def extSimData = new ExtSimDataConnection(extSimAdapter.ref())
         def extSim = new TestSimulation(-1L, Optional.empty())
-        extSim.setAdapterData(extSimData)
+        extSim.setDataConnection(extSimData)
+        extSim.setSetupData(new SetupData(new String[0], null, null))
 
         when:
         extSimData.queueExtMsg(new TerminationMessage(simlulationSuccessful))
@@ -135,9 +139,10 @@ class ExtSimulationSpec extends Specification {
     def "An ExtSimulation should handle unknown messages by throwing an exception"() {
         given:
         def extSimAdapter = testKit.createTestProbe(ControlResponseMessageFromExt)
-        def extSimData = new ExtSimAdapterData(extSimAdapter.ref(), new String[0], null, null)
+        def extSimData = new ExtSimDataConnection(extSimAdapter.ref())
         def extSim = new TestSimulation(-1L, Optional.empty())
-        extSim.setAdapterData(extSimData)
+        extSim.setDataConnection(extSimData)
+        extSim.setSetupData(new SetupData(new String[0], null, null))
 
         when:
         extSimData.queueExtMsg(new UnknownMessage())